+++ nanodjango/constants.py
@@ -1,1 +1,2 @@
 SQLITE_MEMORY = ":memory:"
+SQLITE_TMP = ":tmp:"